requirements.txt
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
Flask==3.1.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
MarkupSafe==3.0.3
pycparser==3.0
pymongo==4.16.0
Werkzeug==3.1.6
python-dotenv==1.0.1